parsers-markdown-extensions: fix incorrect ending of some class blocks
Previously, removing the class indicator split the line in two string
segments, while the whole processing assumes that segments are
equivalent to lines. so the line ending was its own blank segment,
which was considered as an end-of-paragraph. The whole block would
then be stopped unless the following line has the blockquote marker.